Added Realm as a dependency.
Reworked the LongClickListener and debounced it.

Updated the GeofenceManager, which now got a GeofenceRepository for
persistence and a PlayGeofenceManager that will manage the interaction
with the play services.

Added a Realm geofence model to persist the geofences
Added GeofenceData to be the non mutable object used in the application.
Early yet, might change a bit further down.