Upgraded Timber to 4.4.0
Added RxInterop to convert between Rx1 and 2.

Added assertJ to help with tests.

Created the listener for the geofences in the Realm Repositoty.

Created a Mapper class between the domain and model objects.

Created a specific Module for the Realm code, might be a bit overkill, to
be seen.

Created unit tests for the Mapper class.