Upgraded to latest gradle tools and similar.
LocationRequester now relies on an activity as there were problems with
the theme when it was created using the application context.

Moving of classes into different packages.

SelectedGeofence class.
Add, edit and display Geofence will (eventually) all listen
to this class to know which is the currently selected Geofence
and which one that should be displayed.

Moving the display code into a Fragment to be able to use it with other
parts (master / detail ) view for example.

This move is not complete yet! Still lack of testing and edit part is
broken!