Reworked the SelectedGeofence to return a result object. This object hold a boolean informing if the geofence attached is valid or not.
Right now a dummyGeofence is attached if it's not so it can't be null.

Also did something similar with the getGeofence method from the
repository.

The reason for this is that the classes using the methods shouldn't
have to know that a "dummy" geofence can be returned and to check
for that.

Updated and added a few tests for these changes.